refactor(individuo): se extrajo logica comun y se actualizaron tests
