+++ tests/Solver.Tests/Individuos/IndividuoTests.cs
@@ -195,12 +195,12 @@ namespace Solver.Tests.Individuos
                 return 1.0m;
             }
 
-            internal override void Mutar()
+            protected override void MutarAsignaciones()
             {
                 throw new NotImplementedException();
             }
 
-            internal override Individuo Cruzar(Individuo otro)
+            protected override Individuo CrearNuevoIndividuo(List<int> cromosoma)
             {
                 throw new NotImplementedException();
             }
